404 Handling, proposal deletion implemented. UI Tweaks.

+++ managr/managr_react/components/proposals/ProposalTools.jsx
@@ -1,27 +1,61 @@
 import React from 'react';
 
 import { connect } from 'react-redux';
-import { deleteProposal } from '../../actions/ProposalActions';
+import { deleteProposal, resetProposalView } from '../../actions/ProposalActions';
 
 class ProposalTools extends React.Component {
+    componentDidUpdate(prevProps, prevState) {
+        if (this.props.proposal_deleted == true && prevProps.proposal_deleted == false) {
+            // this.props.handleReset();
+            // this.context.router.push('/summary');
+            // What should we do here?
+        }
+    }
+
+    componentWillReceiveProps(next_props) {
+        // If we're changing proposal, force update
+        if(next_props.proposal_uuid != this.props.proposal_uuid) {
+            this.props.resetProposalView();
+        }
+    }
+
     handleDelete(submitEvent) {
-        let session_token = localStorage.getItem("managr_session_token");
-        this.props.deleteProposal(this.props.proposal_uuid, session_token);
+        if(!this.props.proposal_deleted && confirm("Are you sure you want to delete this proposal?")) {
+            let session_token = localStorage.getItem("managr_session_token");
+            this.props.deleteProposal(this.props.proposal_uuid, session_token);
+        }
     }
 
     render () {
         if(this.props.owner == "true") {
+            let deleteSuccess, toolBarClass;
+            if(this.props.proposal_deleted) {
+                deleteSuccess = <p>This proposal has been deleted.</p>;
+                toolBarClass = "alert alert-warning proposal-success";
+            } else {
+                deleteSuccess = <p>This proposal is live on the Managr contractor network.</p>;
+                toolBarClass = "alert alert-success proposal-success";
+            }
+
             return (
-                <div className="alert alert-success proposal-success">
+                <div className={toolBarClass}>
                     <div className="proposal-tool-buttons">
-                        <button className="btn btn-sm btn-default" onClick={this.props.handleClick}>
+                        <button
+                            className="btn btn-sm btn-default"
+                            onClick={this.props.handleClick}
+                            disabled={this.props.proposal_deleted}
+                        >
                             Edit
                         </button>
-                        <button className="btn btn-sm btn-danger" onClick={this.handleDelete.bind(this)}>
+                        <button
+                            className="btn btn-sm btn-danger"
+                            onClick={this.handleDelete.bind(this)}
+                            disabled={this.props.proposal_deleted}
+                        >
                             Delete
                         </button>
                     </div>
-                    <p>This proposal is live on the Managr contractor network.</p>
+                    {deleteSuccess}
                 </div>
             );
         } else {
@@ -39,6 +73,10 @@ class ProposalTools extends React.Component {
     }
 };
 
+ProposalTools.contextTypes = {
+    router: React.PropTypes.object.isRequired
+}
+
 const mapStateToProps = (state) => {
     return {
         proposal_deleted: state.proposal_deleted,
@@ -47,7 +85,8 @@ const mapStateToProps = (state) => {
 
 const mapDispatchToProps = (dispatch) => {
     return {
-        deleteProposal: (proposal_uuid, session_token) => dispatch(deleteProposal(proposal_uuid, session_token))
+        deleteProposal: (proposal_uuid, session_token) => dispatch(deleteProposal(proposal_uuid, session_token)),
+        resetProposalView: () => dispatch(resetProposalView()),
     };
 };
